+++ src/routes/+page.svelte
@@ -146,7 +146,7 @@
 		const unsubUI = uiStore.subscribe((state) => {
 			selectedPlaylistId = state.selectedPlaylistId
 			selectedFolderId = state.selectedFolderId
-			selectedTagIds = state.selectedTagIds
+			selectedTagIds = state.viewFilters[state.activeView].selectedTagIds
 		})
 		const unsubDevices = visibleDevices.subscribe((visibleDevicesList) => {
 			devices = visibleDevicesList
@@ -300,18 +300,19 @@
 		} else if (!restoredFolderId) {
 			// Base-level view: load main data
 			discoveryPlaylistStore.clearReleases()
+			const viewFilters = get(uiStore).viewFilters[view]
 			if (view === 'discovery') {
 				const filter: DiscoveryFilter = {}
-				if (selectedTagIds.length > 0) {
-					filter.tag_ids = selectedTagIds
-					filter.tag_filter_mode = $tagFilterMode
+				if (viewFilters.selectedTagIds.length > 0) {
+					filter.tag_ids = viewFilters.selectedTagIds
+					filter.tag_filter_mode = viewFilters.tagFilterMode
 				}
 				discoveryStore.loadReleases(Object.keys(filter).length > 0 ? filter : undefined)
 			} else {
 				const filter: TrackFilter = {}
-				if (selectedTagIds.length > 0) {
-					filter.tag_ids = selectedTagIds
-					filter.tag_filter_mode = $tagFilterMode
+				if (viewFilters.selectedTagIds.length > 0) {
+					filter.tag_ids = viewFilters.selectedTagIds
+					filter.tag_filter_mode = viewFilters.tagFilterMode
 				}
 				libraryStore.loadTracks(Object.keys(filter).length > 0 ? filter : undefined)
 			}
@@ -435,7 +436,7 @@
 			: (query: string) => libraryStore.setSearch(query)
 	)
 
-	// Active filter tags (for SearchBar tag chips)
+	// Active filter tags
 	const activeFilterTags = $derived.by(() => {
 		if (selectedTagIds.length === 0) return [] as Tag[]
 		const tags: Tag[] = []
@@ -493,10 +494,14 @@
 			{searchValue}
 			{onSearchChange}
 			{activeFilterTags}
+			{tagCategories}
 			tagColors={categoryColors}
 			tagFilterMode={$tagFilterMode}
-			onRemoveTagFilter={(tagId) => tagController.removeTagFilter(tagId)}
-			onClearAllTagFilters={() => tagController.clearTagFilters()}
+			onToggleTagFilter={(tagId) => tagController.selectTag(tagId)}
+			onClearAllTagFilters={() => {
+				tagController.clearTagFilters()
+				if ($activeView === 'discovery' && get(likedOnly)) discoveryStore.toggleLikedFilter()
+			}}
 			onToggleTagFilterMode={() => tagController.toggleTagFilterMode()}
 			isDiscoveryContext={$activeView === 'discovery'}
 			likedOnly={$activeView === 'discovery' ? $likedOnly : false}
@@ -523,10 +528,14 @@
 					searchValue={discoverySearchValue}
 					onSearchChange={(query) => discoveryStore.setSearch(query)}
 					{activeFilterTags}
+					{tagCategories}
 					tagColors={categoryColors}
 					tagFilterMode={$tagFilterMode}
-					onRemoveTagFilter={(tagId) => tagController.removeTagFilter(tagId)}
-					onClearAllTagFilters={() => tagController.clearTagFilters()}
+					onToggleTagFilter={(tagId) => tagController.selectTag(tagId)}
+					onClearAllTagFilters={() => {
+						tagController.clearTagFilters()
+						if (get(likedOnly)) discoveryStore.toggleLikedFilter()
+					}}
 					onToggleTagFilterMode={() => tagController.toggleTagFilterMode()}
 					likedOnly={$likedOnly}
 					onToggleLikedFilter={() => discoveryStore.toggleLikedFilter()}
@@ -558,9 +567,10 @@
 					searchValue={librarySearchValue}
 					onSearchChange={(query) => libraryStore.setSearch(query)}
 					{activeFilterTags}
+					{tagCategories}
 					tagColors={categoryColors}
 					tagFilterMode={$tagFilterMode}
-					onRemoveTagFilter={(tagId) => tagController.removeTagFilter(tagId)}
+					onToggleTagFilter={(tagId) => tagController.selectTag(tagId)}
 					onClearAllTagFilters={() => tagController.clearTagFilters()}
 					onToggleTagFilterMode={() => tagController.toggleTagFilterMode()}
 					onSelectionChange={trackController.handleSelectionChange}
@@ -592,10 +602,14 @@
 			searchValue={discoverySearchValue}
 			onSearchChange={(query) => discoveryStore.setSearch(query)}
 			{activeFilterTags}
+			{tagCategories}
 			tagColors={categoryColors}
 			tagFilterMode={$tagFilterMode}
-			onRemoveTagFilter={(tagId) => tagController.removeTagFilter(tagId)}
-			onClearAllTagFilters={() => tagController.clearTagFilters()}
+			onToggleTagFilter={(tagId) => tagController.selectTag(tagId)}
+			onClearAllTagFilters={() => {
+				tagController.clearTagFilters()
+				if (get(likedOnly)) discoveryStore.toggleLikedFilter()
+			}}
 			onToggleTagFilterMode={() => tagController.toggleTagFilterMode()}
 			likedOnly={$likedOnly}
 			onToggleLikedFilter={() => discoveryStore.toggleLikedFilter()}
@@ -630,9 +644,10 @@
 			searchValue={librarySearchValue}
 			onSearchChange={(query) => libraryStore.setSearch(query)}
 			{activeFilterTags}
+			{tagCategories}
 			tagColors={categoryColors}
 			tagFilterMode={$tagFilterMode}
-			onRemoveTagFilter={(tagId) => tagController.removeTagFilter(tagId)}
+			onToggleTagFilter={(tagId) => tagController.selectTag(tagId)}
 			onClearAllTagFilters={() => tagController.clearTagFilters()}
 			onToggleTagFilterMode={() => tagController.toggleTagFilterMode()}
 			onSelectionChange={trackController.handleSelectionChange}